requirements.txt
beautifulsoup4
requests
pandas
tqdm
langdetect
nltk
rich-logging
scikit-learn
matplotlib
gensim
openai
mmh3
networkx
PyYAML
wordcloud
transformers
torch
